Rest Template

+++ hr-payroll/src/main/java/br/com/hr_payroll/resources/PaymentResource.java
@@ -18,7 +18,7 @@ public class PaymentResource {
 	private PaymentService service;
 	
 	@GetMapping("/{workerId}/days/{days}")
-	public ResponseEntity<Payment> getPayment(@PathVariable Double workerId, @PathVariable Integer days){
+	public ResponseEntity<Payment> getPayment(@PathVariable Long workerId, @PathVariable Integer days){
 		Payment payment = service.getPayment(workerId, days);
 		return ResponseEntity.ok(payment);
 	}
